# docs/api/pms-openapi.json
{"openapi":"3.1.0","info":{"title":"SmartHotel PMS Core API","description":"Public booking and staff admin API.\n\nError contract: RFC 7807 `application/problem+json` (`https://smarthotel/problems/…`). Full catalog: `docs/api/pms-api.md`.\n","contact":{"name":"SmartHotel-PMS","url":"https://github.com"},"license":{"name":"MIT"},"version":"v1"},"servers":[{"url":"/","description":"Current host"}],"tags":[{"name":"Auth","description":"Staff login and token refresh"},{"name":"Availability","description":"Public inventory search"},{"name":"Reservations","description":"Public booking create / lookup / cancel"},{"name":"Admin — Room types","description":"ADMIN only"},{"name":"Admin — Rooms","description":"ADMIN only"},{"name":"Admin — Guests","description":"ADMIN or RECEPTIONIST"},{"name":"Admin — Reservations","description":"ADMIN or RECEPTIONIST"},{"name":"Admin — Rate calendar","description":"View both roles; MANUAL override ADMIN"},{"name":"Admin — Rate plans","description":"ADMIN or RECEPTIONIST"},{"name":"Admin — Dashboard","description":"ADMIN or RECEPTIONIST"},{"name":"Admin — Pricing proxies","description":"Phase 8 stubs until pricing-service is wired"}],"paths":{"/api/v1/admin/rooms/{id}":{"get":{"tags":["Admin — Rooms"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomResponse"}}}}},"security":[{"bearer-jwt":[]}]},"put":{"tags":["Admin — Rooms"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoomRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/room-types/{id}":{"get":{"tags":["Admin — Room types"],"operationId":"get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomTypeResponse"}}}}},"security":[{"bearer-jwt":[]}]},"put":{"tags":["Admin — Room types"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoomTypeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomTypeResponse"}}}}},"security":[{"bearer-jwt":[]}]},"delete":{"tags":["Admin — Room types"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/rate-calendar/{roomTypeCode}/{date}":{"put":{"tags":["Admin — Rate calendar"],"operationId":"putManual","parameters":[{"name":"roomTypeCode","in":"path","required":true,"schema":{"type":"string"}},{"name":"date","in":"path","required":true,"schema":{"type":"string","format":"date"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManualRateOverrideRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RateCalendarDayResponse"}}}}},"security":[{"bearer-jwt":[]}]},"delete":{"tags":["Admin — Rate calendar"],"operationId":"deleteManual","parameters":[{"name":"roomTypeCode","in":"path","required":true,"schema":{"type":"string"}},{"name":"date","in":"path","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"204":{"description":"No Content"}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/guests/{id}":{"get":{"tags":["Admin — Guests"],"operationId":"get_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GuestResponse"}}}}},"security":[{"bearer-jwt":[]}]},"put":{"tags":["Admin — Guests"],"operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGuestRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GuestResponse"}}}}},"security":[{"bearer-jwt":[]}]},"delete":{"tags":["Admin — Guests"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}},"security":[{"bearer-jwt":[]}]}},"/api/v1/reservations":{"post":{"tags":["Reservations"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReservationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}}}},"/api/v1/reservations/{code}/cancel":{"post":{"tags":["Reservations"],"operationId":"cancel","parameters":[{"name":"code","in":"path","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["Auth"],"summary":"Silent refresh","description":"Reads the httpOnly refresh cookie (default name pms_refresh), rotates tokens, and returns a new access JWT. No request body.","operationId":"refresh","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["Auth"],"summary":"Staff logout","description":"Clears the refresh cookie.","operationId":"logout","responses":{"200":{"description":"OK"}}}},"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"Staff login","description":"Returns an access JWT in the body and sets the refresh JWT as an httpOnly; SameSite=Strict cookie (ADR-0014).","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/v1/admin/rooms":{"get":{"tags":["Admin — Rooms"],"operationId":"list","parameters":[{"name":"roomTypeId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["AVAILABLE","OUT_OF_SERVICE"]}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseRoomResponse"}}}}},"security":[{"bearer-jwt":[]}]},"post":{"tags":["Admin — Rooms"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoomRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/room-types":{"get":{"tags":["Admin — Room types"],"operationId":"list_1","parameters":[{"name":"active","in":"query","required":false,"schema":{"type":"boolean"}},{"name":"query","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseRoomTypeResponse"}}}}},"security":[{"bearer-jwt":[]}]},"post":{"tags":["Admin — Room types"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoomTypeRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomTypeResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/reservations":{"get":{"tags":["Admin — Reservations"],"operationId":"list_2","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["CONFIRMED","CHECKED_IN","CHECKED_OUT","CANCELLED","NO_SHOW"]}},{"name":"from","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"to","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"query","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseAdminReservationSummary"}}}}},"security":[{"bearer-jwt":[]}]},"post":{"tags":["Admin — Reservations"],"operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReservationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/reservations/{id}/no-show":{"post":{"tags":["Admin — Reservations"],"operationId":"noShow","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/reservations/{id}/check-out":{"post":{"tags":["Admin — Reservations"],"operationId":"checkOut","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/reservations/{id}/check-in":{"post":{"tags":["Admin — Reservations"],"operationId":"checkIn","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/reservations/{id}/cancel":{"post":{"tags":["Admin — Reservations"],"operationId":"cancel_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/pricing/refresh":{"post":{"tags":["Admin — Pricing proxies"],"operationId":"refresh_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PricingRefreshResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/guests":{"get":{"tags":["Admin — Guests"],"operationId":"list_3","parameters":[{"name":"query","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseGuestResponse"}}}}},"security":[{"bearer-jwt":[]}]},"post":{"tags":["Admin — Guests"],"operationId":"create_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGuestRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GuestResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/reservations/lookup":{"get":{"tags":["Reservations"],"operationId":"lookup","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}}}},"/api/v1/availability":{"get":{"tags":["Availability"],"operationId":"search","parameters":[{"name":"checkIn","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"checkOut","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"guests","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AvailabilityResponse"}}}}}}},"/api/v1/admin/reservations/{id}":{"get":{"tags":["Admin — Reservations"],"operationId":"get_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/rate-plans":{"get":{"tags":["Admin — Rate plans"],"operationId":"list_4","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RatePlanResponse"}}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/rate-calendar":{"get":{"tags":["Admin — Rate calendar"],"operationId":"get_4","parameters":[{"name":"roomTypeCode","in":"query","required":true,"schema":{"type":"string"}},{"name":"from","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"to","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RateCalendarResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/pricing/events":{"get":{"tags":["Admin — Pricing proxies"],"operationId":"events","parameters":[{"name":"from","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"to","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PricingEventResponse"}}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/pricing/demand-indicators":{"get":{"tags":["Admin — Pricing proxies"],"operationId":"demandIndicators","parameters":[{"name":"from","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"to","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DemandIndicatorPoint"}}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/dashboard/timeseries":{"get":{"tags":["Admin — Dashboard"],"operationId":"timeseries","parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":30}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardTimeseriesResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/admin/dashboard/kpis":{"get":{"tags":["Admin — Dashboard"],"operationId":"kpis","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardKpisResponse"}}}}},"security":[{"bearer-jwt":[]}]}}},"components":{"schemas":{"UpdateRoomRequest":{"type":"object","properties":{"roomNumber":{"type":"string","maxLength":10,"minLength":0},"roomTypeId":{"type":"integer","format":"int64"},"floor":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["AVAILABLE","OUT_OF_SERVICE"]},"notes":{"type":"string"}},"required":["roomTypeId","status"]},"RoomResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"roomNumber":{"type":"string"},"roomTypeId":{"type":"integer","format":"int64"},"roomTypeCode":{"type":"string"},"floor":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["AVAILABLE","OUT_OF_SERVICE"]},"notes":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"UpdateRoomTypeRequest":{"type":"object","properties":{"code":{"type":"string","maxLength":30,"minLength":0},"name":{"type":"string","maxLength":100,"minLength":0},"description":{"type":"string"},"capacity":{"type":"integer","format":"int32","maximum":10,"minimum":1},"basePrice":{"type":"number","minimum":0.01},"minPrice":{"type":"number","minimum":0.01},"maxPrice":{"type":"number","minimum":0.01},"amenities":{"type":"array","items":{"type":"string"}},"active":{"type":"boolean"},"validPriceBand":{"type":"boolean"}},"required":["active","basePrice","capacity","maxPrice","minPrice"]},"RoomTypeResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"capacity":{"type":"integer","format":"int32"},"basePrice":{"type":"number"},"minPrice":{"type":"number"},"maxPrice":{"type":"number"},"currency":{"type":"string"},"amenities":{"type":"array","items":{"type":"string"}},"active":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ManualRateOverrideRequest":{"type":"object","properties":{"price":{"type":"number","minimum":0.00}},"required":["price"]},"RateCalendarDayResponse":{"type":"object","properties":{"date":{"type":"string","format":"date"},"price":{"type":"number"},"source":{"type":"string","enum":["ML_MODEL","MANUAL","BASE","BASE_FALLBACK"]},"demandIndicator":{"type":"integer","format":"int32"},"fromCalendar":{"type":"boolean"}}},"UpdateGuestRequest":{"type":"object","properties":{"firstName":{"type":"string","maxLength":80,"minLength":0},"lastName":{"type":"string","maxLength":80,"minLength":0},"email":{"type":"string","maxLength":255,"minLength":0},"phone":{"type":"string","maxLength":30,"minLength":0}}},"GuestResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateReservationRequest":{"type":"object","properties":{"roomTypeCode":{"type":"string","maxLength":30,"minLength":0},"ratePlanCode":{"type":"string","maxLength":30,"minLength":0},"checkIn":{"type":"string","format":"date"},"checkOut":{"type":"string","format":"date"},"adults":{"type":"integer","format":"int32","maximum":10,"minimum":1},"guest":{"$ref":"#/components/schemas/GuestUpsertRequest"},"validStay":{"type":"boolean"}},"required":["adults","checkIn","checkOut","guest"]},"GuestUpsertRequest":{"type":"object","properties":{"firstName":{"type":"string","maxLength":80,"minLength":0},"lastName":{"type":"string","maxLength":80,"minLength":0},"email":{"type":"string","maxLength":255,"minLength":0},"phone":{"type":"string","maxLength":30,"minLength":0}}},"PriceBreakdownLine":{"type":"object","properties":{"date":{"type":"string","format":"date"},"price":{"type":"number"},"barPrice":{"type":"number"},"priceSource":{"type":"string","enum":["ML_MODEL","MANUAL","BASE","BASE_FALLBACK"]}}},"RatePlanSummary":{"type":"object","properties":{"code":{"type":"string"},"refundable":{"type":"boolean"},"breakfastIncluded":{"type":"boolean"}}},"ReservationResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"confirmationCode":{"type":"string"},"status":{"type":"string","enum":["CONFIRMED","CHECKED_IN","CHECKED_OUT","CANCELLED","NO_SHOW"]},"roomType":{"type":"string"},"roomNumber":{"type":"string"},"ratePlan":{"$ref":"#/components/schemas/RatePlanSummary"},"checkIn":{"type":"string","format":"date"},"checkOut":{"type":"string","format":"date"},"adults":{"type":"integer","format":"int32"},"totalPrice":{"type":"number"},"currency":{"type":"string"},"priceBreakdown":{"type":"array","items":{"$ref":"#/components/schemas/PriceBreakdownLine"}},"source":{"type":"string","enum":["WEB","ADMIN"]}}},"TokenResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"tokenType":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"},"email":{"type":"string"},"fullName":{"type":"string"},"role":{"type":"string","enum":["ADMIN","RECEPTIONIST"]}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}}},"CreateRoomRequest":{"type":"object","properties":{"roomNumber":{"type":"string","maxLength":10,"minLength":0},"roomTypeId":{"type":"integer","format":"int64"},"floor":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["AVAILABLE","OUT_OF_SERVICE"]},"notes":{"type":"string"}},"required":["roomTypeId"]},"CreateRoomTypeRequest":{"type":"object","properties":{"code":{"type":"string","maxLength":30,"minLength":0},"name":{"type":"string","maxLength":100,"minLength":0},"description":{"type":"string"},"capacity":{"type":"integer","format":"int32","maximum":10,"minimum":1},"basePrice":{"type":"number","minimum":0.01},"minPrice":{"type":"number","minimum":0.01},"maxPrice":{"type":"number","minimum":0.01},"amenities":{"type":"array","items":{"type":"string"}},"active":{"type":"boolean"},"validPriceBand":{"type":"boolean"}},"required":["basePrice","capacity","maxPrice","minPrice"]},"PricingRefreshResponse":{"type":"object","properties":{"status":{"type":"string"},"message":{"type":"string"},"daysUpdated":{"type":"integer","format":"int32"}}},"CreateGuestRequest":{"type":"object","properties":{"firstName":{"type":"string","maxLength":80,"minLength":0},"lastName":{"type":"string","maxLength":80,"minLength":0},"email":{"type":"string","maxLength":255,"minLength":0},"phone":{"type":"string","maxLength":30,"minLength":0}}},"AvailabilityResponse":{"type":"object","properties":{"checkIn":{"type":"string","format":"date"},"checkOut":{"type":"string","format":"date"},"guests":{"type":"integer","format":"int32"},"currency":{"type":"string"},"roomTypes":{"type":"array","items":{"$ref":"#/components/schemas/AvailableRoomTypeAvailability"}}}},"AvailableRoomTypeAvailability":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"},"capacity":{"type":"integer","format":"int32"},"roomsLeft":{"type":"integer","format":"int64"},"nights":{"type":"array","items":{"$ref":"#/components/schemas/NightAvailability"}},"ratePlans":{"type":"array","items":{"$ref":"#/components/schemas/RatePlanQuote"}}}},"NightAvailability":{"type":"object","properties":{"date":{"type":"string","format":"date"},"bar":{"type":"number"},"priceSource":{"type":"string","enum":["ML_MODEL","MANUAL","BASE","BASE_FALLBACK"]}}},"RatePlanQuote":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"},"refundable":{"type":"boolean"},"breakfastIncluded":{"type":"boolean"},"totalPrice":{"type":"number"}}},"PageMeta":{"type":"object","properties":{"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"}}},"PageResponseRoomResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/RoomResponse"}},"page":{"$ref":"#/components/schemas/PageMeta"}}},"PageResponseRoomTypeResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/RoomTypeResponse"}},"page":{"$ref":"#/components/schemas/PageMeta"}}},"AdminReservationSummary":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"confirmationCode":{"type":"string"},"status":{"type":"string","enum":["CONFIRMED","CHECKED_IN","CHECKED_OUT","CANCELLED","NO_SHOW"]},"roomType":{"type":"string"},"roomNumber":{"type":"string"},"guestName":{"type":"string"},"guestEmail":{"type":"string"},"checkIn":{"type":"string","format":"date"},"checkOut":{"type":"string","format":"date"},"adults":{"type":"integer","format":"int32"},"totalPrice":{"type":"number"},"currency":{"type":"string"},"source":{"type":"string","enum":["WEB","ADMIN"]}}},"PageResponseAdminReservationSummary":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/AdminReservationSummary"}},"page":{"$ref":"#/components/schemas/PageMeta"}}},"RatePlanResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"refundable":{"type":"boolean"},"breakfastIncluded":{"type":"boolean"},"priceModifier":{"type":"number"},"active":{"type":"boolean"},"sortOrder":{"type":"integer","format":"int32"}}},"RateCalendarResponse":{"type":"object","properties":{"roomTypeCode":{"type":"string"},"days":{"type":"array","items":{"$ref":"#/components/schemas/RateCalendarDayResponse"}}}},"PricingEventResponse":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"impactScore":{"type":"number"},"confidence":{"type":"number"},"rationale":{"type":"string"},"source":{"type":"string"}}},"DemandIndicatorPoint":{"type":"object","properties":{"date":{"type":"string","format":"date"},"demandIndicator":{"type":"integer","format":"int32"}}},"PageResponseGuestResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/GuestResponse"}},"page":{"$ref":"#/components/schemas/PageMeta"}}},"DashboardDayPoint":{"type":"object","properties":{"date":{"type":"string","format":"date"},"occupancy":{"type":"number"},"adr":{"type":"number"},"roomsOccupied":{"type":"integer","format":"int64"},"roomsSellable":{"type":"integer","format":"int64"}}},"DashboardTimeseriesResponse":{"type":"object","properties":{"from":{"type":"string","format":"date"},"to":{"type":"string","format":"date"},"currency":{"type":"string"},"points":{"type":"array","items":{"$ref":"#/components/schemas/DashboardDayPoint"}}}},"DashboardKpisResponse":{"type":"object","properties":{"businessDate":{"type":"string","format":"date"},"occupancyToday":{"type":"number"},"roomsOccupiedToday":{"type":"integer","format":"int64"},"roomsSellable":{"type":"integer","format":"int64"},"arrivalsToday":{"type":"integer","format":"int64"},"departuresToday":{"type":"integer","format":"int64"},"mtdRevenue":{"type":"number"},"currency":{"type":"string"}}}},"securitySchemes":{"bearer-jwt":{"type":"http","description":"Staff access token from POST /api/v1/auth/login (HS256, ~60 min). Refresh tokens are not accepted here.","name":"bearer-jwt","scheme":"bearer","bearerFormat":"JWT"}}}}